// application/banca-edge-ui/src/app/complaints/model/complaint-info.model.ts
export class ComplaintModel {
    policyNumber: string;
    detailsCorrection: string;
    customerMailId: string;
    complaintAgainst: string;
    complaintMode: string;
    complaintNature: string;
    uploadDocuments: string;
    status: string;
    closureRemarks?: string;
    openDate: Date;
    closeDate?: Date;

}